Move resource tracking/locking back from Queue to futures
